bugfix: use the actual expiration on dealer timeout entries
we were using a discrete number as the timeout entry and on each
`timeout_step` this value was decremented. when it reached zero, the
system assumed that entry expired. this mechanism was flawed because
at the time the timeout entry is created the elapsed time until the
next `timeout_step` execute might actually be really small. thus the
entry times out almost immediately.

we could fix by using a greater initial value, demanding two or more
runs between considering an entry expired. instead we decided to use
the deadline itself and the Timeout module which give us a precise
mechanism.