improve logger
the purpose of this change is to allow suppressing logging messages or
changing the library log level. we've have to build this mechanism in
the library as elixir has no mechanism that allow configuring this per
application.

the default implementation uses `Logger.log/2` as the backend but you
can use any 2-arity function that takes the log level and the
message (as a function).

it is also possible to disable the logging completely by using:
`Application.set_env(:liblink, :logger, nil)`

incidentally we also change the format of the response when an
exception occurs. the stacktrace is no longer send to the peer, only
the exception.